tofu: Remove GraphNodeDynamicExpandable
A long time ago in a predecessor codebase far, far away this package used
a whole lot of inversion-of-control techniques, where instead of just
calling a function to do a thing we'd instead call a function to get a
data structure describing a set of things to do and then some caller
would arrange for those things to actually get done.

GraphNodeExecutable was a first step towards moving away from that toward
writing "normal Go code" that just runs to completion. It was successful
in causing a bunch of things that were previously spread awkwardly around
the codebase into straight-though code that a reader can actually follow
sequentially to understand what it does and what order it does it in.

At the time we'd wanted to also remove GraphNodeDynamicExpandable, because
it's also a rather needless inversion of control: in theory it allows a
graph node to produce another "subgraph" that gets walked as part of the
visit to the parent node, but in practice all of these "subgraphs" were
really just flat sets of functions to call but obscured behind the
tofu.Graph type and executed using the graph walker.

As a further small step towards making this code more "normal", this
removes the GraphNodeDynamicExpandable interface altogether and instead
incorporates the same behavior into the GraphNodeExecutable
implementations.

The graph walker would always deal with DynamicExpand immediately after
Execute anyway, and so performing the same actions at the end of Execute
is functionally equivalent.

To avoid changing too much at once this does still retain the ceremony of
building a bunch of "graph nodes", but then executes them using a simple
helper function tofu.executeGraphNodes instead of using the full graph
walker machinery. The goal of this initial commit is just to demonstrate
that we can get equivalent behavior despite the removal of this extra
abstraction, while keeping the changes relatively small.

Most of these DynamicExpand functions had previously been simplified to
directly construct a graph instead of using the "graph transformer"
machinery, in preparation for a change like this. Those therefore can now
just produce a flat sequence of GraphNodeExecutable and not bother
constructing a tofu.Graph at all. nodeExpandPlannableResource is more
change-resistant due to its greater complexity and so it retains the use
of the graph builders but now just throws away the graph container and
executes the nodes within it directly.

Future commits should gradually simplify this further, ideally avoiding
any need to construct separate "graph node" objects altogether and instead
just using normal function calls with arguments so that the overall data
flow and control flow is easier to follow. The executeGraphNodes helper
illustrates one simple pattern for still performing those function calls
concurrently, but without all of the dependency tracking of the graph
walker that isn't needed when you don't actually have any dependencies.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>