tofu: EvalContext.PerformIO method
This commit is an interim step toward a future change to the way we handle
the concurrency limit for I/O operations during a graph walk.

Today we have a centrally-managed semaphore that is encapsulated in the
graph walker so that the limit is imposed on the number of graph nodes
that can have their Execute methods running at once, rather than on the
I/O operations themselves.

That design requires a rather confusing inversion-of-control approach that
we started to move away from in an earlier commit, but that came at the
expense of no longer respecting the concurrency limit for leaf operations
that now all happen together in a single graph node. This particular commit
doesn't correct that regression, but it sets us up to correct it in a
forthcoming commit that will make the concurrency handling explicit in
the graph node execute functions rather than something hidden/automatic
in the graph walker.

Along with just making things more explicit and thus hopefully easier to
understand and maintain, this new design will also give more flexibility
to individual graph node implementations on how they choose to handle their
execution, such as finally meeting the long-standing request to force
instances of a resource to be visited sequentially instead of concurrently,
should we choose to allow that in future.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>