tofu: Concurrency limits now handled explicitly in node Execute
Previously we implemented I/O concurrency limits very coarsely in the
central graph walker mechanism, which forced every node type to be treated
the same way in return for allowing graph node implementors to not really
think about the concurrency limit at all.

However, that approach required a confusing inversion-of-control model that
made the control flow quite hard to navigate. This has been a recurring
challenge for most newcomers to the codebase, and so we've gradually been
reducing the inversion-of-control-shaped approaches in favor of writing
more straightforward sequential code that is hopefully easier for a
newcomer to follow and understand how the language runtime handles each
different type of object.

This commit continues that journey by removing the semaphore responsibility
from the graph walker entirely, and instead expecting individual graph
node implementations to call EvalContext.PerformIO any time they are
performing some work that should be subjected to the concurrency limit.

Since most of our graph node types that perform I/O are still using the
tofu.executeGraphNodes helper shim to continue using something like their
previous inversion-of-control approach, that helper is now effectively
restoring the previous concurrency limit functionality for those node
types without any individual effort. Future work to refactor each
individual set of node types will deal with figuring out where is best
to call evalCtx.PerformIO inline.

However, a notable exception is that during the apply phase individual
resource instances get directly added to the graph based on their
appearance in the saved plan, and so NodeApplyableResourceInstance gets
to be the first graph node type to handle its own I/O concurrency rather
than expecting its caller to handle it. It's still handled rather coarsely
for now to limit the size of the change, but we can improve on that in
future commits as we get more experience with what seems like the most
appropriate granularity for EvalContext.PerformIO calls.

Administratively, this also includes some work in our ongoing effort to
switch to using the symbol name "evalCtx" for variables of type
tofu.EvalContext, to make room for using "ctx" to refer to context.Context
when we plumb that into GraphNodeExecutable in future.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>