tofu: Simplify output value evaluation
Output value evaluation was previously split over three graph node types
due it having previously been written for the now-removed
GraphNodeDynamicExpandable abstraction.

The split into three graph node types doesn't really have any benefit and
just made the code harder to follow, so this consolidates nodeExpandOutput,
NodeApplyableOutput, and NodeDestroyableOutput into a single
nodeOutputValue that deals with both the multi-instance expansion and the
evaluation of each discovered instance.

The new node type has an unexported name as part of our ongoing effort to
un-export all of the implementation details of this package, so that its
public API is easier to follow in the package documentation.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>